requirements.txt
coverage==6.0.2
Pillow==9.0.0
praw==7.5.0
pytest==6.2.5
pytest-cov==3.0.0
python-telegram-bot==13.9
requests==2.26.0
